backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.3.0
opencv-python-headless==4.9.0.80
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
reportlab==4.1.0